src/evaluation_metrics/requirements.txt
langchain
langchain_openai
python-dotenv
typing